+++ src/main/java/org/ryjan/telegram/interfaces/UserDAO.java
@@ -4,8 +4,8 @@ import org.ryjan.telegram.database.BankDatabase;
 import org.ryjan.telegram.database.UserDatabase;
 
 public interface UserDAO {
-    public UserDatabase findById(int id);
-    public BankDatabase findBankById(int id);
+    public UserDatabase findById(long id);
+    public BankDatabase findBankById(long id);
     public void save(UserDatabase userDatabase);
     public void update(UserDatabase userDatabase);
     public void delete(UserDatabase userDatabase);
